look for $FISH_CONFIG_HOME ahead of XDG
Adds a layer of flexibility for the location of config.fish,
functions, etc. by allowing the user to specify an explicit fish
directory; not necessarily one named 'fish'.

If `$FISH_CONFIG_HOME` is present, it will be used. Otherwise the
location will fallback on `$XDG_CONFIG_HOME/fish`, and
`$HOME/.config/fish` in turn - as prior to this commit.

This commit implements #3030.